% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggauge.R
\name{gggauge}
\alias{gggauge}
\title{Title}
\usage{
gggauge(
  data = NULL,
  type = c("needle", "level"),
  animate = FALSE,
  data_labels = NULL,
  gauge_width = 1/3,
  gauge_colour = "grey30",
  gauge_fill = "transparent",
  plot_ticks = TRUE,
  tick_length = 0.05,
  tick_colour = "grey70",
  plot_tick_labels = TRUE,
  tick_label_colour = "grey50",
  tick_label_size = 4,
  rm_obscured_labels = FALSE,
  limits = c(NA, NA),
  num_breaks = 11,
  breaks = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
